+++ src/legacy/unit.go
@@ -24,7 +24,7 @@ func Nox_xxx_mobInformOwnerHP_4EE4C0(a1 *server.Object) {
 	nox_xxx_mobInformOwnerHP_4EE4C0(asObjectC(a1))
 }
 func Nox_xxx_protectMana_56F9E0(a1 int, a2 int16) {
-	nox_xxx_protectMana_56F9E0(int32(a1), a2)
+	nox_xxx_protectMana_56F9E0(uint32(int32(a1)), a2)
 }
 func Nox_xxx_monsterWalkTo_514110(a1 *server.Object, a2 float32, a3 float32) {
 	nox_xxx_monsterWalkTo_514110(asObjectC(a1), a2, a3)
